add minor query optimization and add option to keep existing users

+++ api_yamdb/reviews/management/commands/import_4_users.py
@@ -12,12 +12,24 @@ class Command(BaseCommand):
     2. Импортирует данные из указанного csv-файла, подставляет их в поля
        модели и заполняет базу новыми объектами.
     Запуск команды: python3 manage.py import_4_users
+
+    Нумерация пользователей в csv начинается с 100, поэтому, если нет желания
+    удалять уже занесенных в базу пользователей (и их меньше 100), то можно
+    запустить команду с опцией --do-not-delete-users
     """
 
-    def handle(self, *args, **kwargs):
-        User.objects.all().delete()
-        csv_file_path = settings.BASE_DIR / 'static/data/users.csv'
+    def add_arguments(self, parser):
+        parser.add_argument(
+            '--do-not-delete-users',
+            action='store_true',
+            help='Do not delete existing users'
+        )
 
+    def handle(self, *args, **kwargs):
+        if not kwargs['do_not_delete_users']:
+            User.objects.all().delete()
+        csv_file_path = settings.BASE_DIR / 'static/data/users.csv'
+        users = []
         with open(csv_file_path, 'r') as file:
             reader = csv.DictReader(file)
             for row in reader:
@@ -27,4 +39,5 @@ class Command(BaseCommand):
                 new_object.email = row['email']
                 new_object.role = row['role']
                 new_object.bio = row['bio']
-                new_object.save()
+                users.append(new_object)
+        User.objects.bulk_create(users)
